chore: move function code to utils
