+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "twikoo",
-  "version": "1.5.11",
+  "version": "1.6.0-beta.1",
   "description": "A simple comment system.",
   "keywords": [
     "twikoojs",
