+++ src/client/version.js
@@ -1,3 +1,3 @@
-const version = '1.5.11'
+const version = '1.6.0-beta.1'
 
 export { version }
