+++ src/server/function/twikoo/utils/spam.js
@@ -1,0 +1,57 @@
+const {
+  AkismetClient,
+  CryptoJS,
+  tencentcloud
+} = require('./lib')
+
+module.exports = {
+  // 后垃圾评论检测
+  async postCheckSpam (comment, config) {
+    try {
+      let isSpam
+      if (comment.isSpam) {
+        // 预检测没过的，就不再检测了
+        isSpam = true
+      } else if (config.QCLOUD_SECRET_ID && config.QCLOUD_SECRET_KEY) {
+        // 腾讯云内容安全
+        const client = new tencentcloud.tms.v20200713.Client({
+          credential: { secretId: config.QCLOUD_SECRET_ID, secretKey: config.QCLOUD_SECRET_KEY },
+          region: 'ap-shanghai',
+          profile: { httpProfile: { endpoint: 'tms.tencentcloudapi.com' } }
+        })
+        const checkResult = await client.TextModeration({
+          Content: CryptoJS.enc.Base64.stringify(CryptoJS.enc.Utf8.parse(comment.comment)),
+          Device: { IP: comment.ip },
+          User: { Nickname: comment.nick }
+        })
+        console.log('腾讯云返回结果：', checkResult)
+        isSpam = checkResult.EvilFlag !== 0
+      } else if (config.AKISMET_KEY) {
+        // Akismet
+        const akismetClient = new AkismetClient({
+          key: config.AKISMET_KEY,
+          blog: config.SITE_URL
+        })
+        const isValid = await akismetClient.verifyKey()
+        if (!isValid) {
+          console.log('Akismet key 不可用：', config.AKISMET_KEY)
+          return
+        }
+        isSpam = await akismetClient.checkSpam({
+          user_ip: comment.ip,
+          user_agent: comment.ua,
+          permalink: comment.href,
+          comment_type: comment.rid ? 'reply' : 'comment',
+          comment_author: comment.nick,
+          comment_author_email: comment.mail,
+          comment_author_url: comment.link,
+          comment_content: comment.comment
+        })
+      }
+      console.log('垃圾评论检测结果：', isSpam)
+      return isSpam
+    } catch (err) {
+      console.error('垃圾评论检测异常：', err)
+    }
+  }
+}
