+++ src/server/self-hosted/package.json
@@ -1,17 +1,17 @@
 {
-  "name": "twikoo-vercel",
+  "name": "twikoo-server",
   "version": "1.6.0-beta.1",
   "description": "A simple comment system.",
   "author": "imaegoo <hello@imaegoo.com> (https://github.com/imaegoo)",
   "license": "MIT",
-  "main": "api/index.js",
+  "main": "index.js",
   "repository": {
     "type": "git",
     "url": "https://github.com/imaegoo/twikoo.git"
   },
   "homepage": "https://twikoo.js.org",
   "dependencies": {
-    "mongodb": "^3.6.3",
+    "rxdb": "^12.7.16",
     "twikoo-func": "1.6.0-beta.1",
     "uuid": "^8.3.2"
   }
